requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
